Fixed sessions widget JS error for novnc connections
